test(api): speed up flaky document-service integration tests
Use minimal article/category fixtures and split slow suites so afterAll
teardown stays under the 60s Jest hook timeout on CI Postgres runners.
